Added start, end dates to timecard; reworked views
The fields start_date and end_date were added to the timecard model,
which represent the range in which eventdates must fall to be on that
timecard. Some validation code was added to check this, and other existing
validation code was modified.

Most of the views for timecards and timecard_entries were changed to
beter follow DRY and make the views easier to use and more consistent.
Part of the view change is that timecards/index is intended for financiers
and not general members.