Got `rails s` to work on development machine
rmail is a very old gem. I'm not exactly sure what it was being used for because Ed commented stuff out but we will see.

Also deleted all migrations (because they are not canonical and not necessary at this stage of development) and dumped a database schema from psuedo-production which has been judged to be canonical.