Possibly fixed email pulling
As I'm not actually sure as to what the problem with email pulling is, this is just a preliminary fix. There were two things that I changed that may have fixed this. I changed the IMAP server from Cyrus to Google, since CMU switched email providers months ago, and I'm not sure if Cyrus actually works anymore. Second, the app now checks the email account's inbox rather than a special folder called "user.abtech.todo".

I also added some nice pagination for the email list view, and fixed a couple of problems regarding unfiling an email. Emails can be associated with an event, but the database schema says that the event_id column cannot be null, thus forcing all new emails to be associated with the event with id 0. Because there just happens to be no event with such an id, this technically works, and email.event returns nil. However, unfiling sets the email's event to nil, which is not allowed. To fix things up, I allowed event_id to be nil and now unfiled emails are always nil, not zero.