Removed accounts/view
As far as I can tell, this action has not worked in a very long time, and is all but inaccessible from the app (other than a link from journal/list which itself is not easy to access). It appears to show a list of journals involving that account, in what appears to be an actual accounting system as it appears journals at that point had a credit account and a debit account. I do not think this is being used anymore, so it is being removed.