Protected events controller
Exec can manage all aspects of an event. An event's TIC can manage all aspects of their event. Someone who has a run position on an event can upload/delete attachments, and modify the notes for an event, as well as remove themself from the run position, and manage the run positions of those who are their assistants (aSM for SM, aME for ME, and so on).

Note that at this point we are disallowing anybody who isn't a tracker dev from deleting events.

refs #91