Fixed pulling of multipart emails
Current behavior is to pick the text part of the email; if none exists, take the HTML part and strip the tags. If there is also no HTML part, then we have a really messed up email, and we show an error telling the user to go to the webmail.