Fixed new timecard caching bug
The cause of this long-standing issue was the validation in timecard_entry itself. In production mode, the array of timecards to validate against is evaluated once, when the class is loaded, which is 1) why the issue went away after restarting Apache, and 2) why it was so difficult to reproduce outside of production. The problem is solved by putting the array in a lambda which is then evaluated every time the validation is called instead of only when the class is loaded.

refs #230