Added notes to eventdates
Also slightly redesigned the events show page. Each eventdate now has its own block, rather than all of the schedules being shoved into one block. An eventdate with notes will show the notes below the schedule (within the same block). The notes for the entire event is in a block on the right now instead of on the left.

refs #215