Changed new member workflow
You can no longer set a password when creating a new member. Instead, an unusuable and hidden random password is assigned to a new member so that they cannot log into the account using a password. If the member has an @andrew.cmu.edu email, they can log in normally using Shibboleth, and once logged in, they may set a password if they wish to. If the member does not have an @andrew.cmu.edu email, a password reset token is emailed to the new member allowing them to set a password.

refs #256