Fixed bug with emoji in email
This mainly involved converting the database to use the utf8mb4
character set. More information on this fix can be found here:
https://mathiasbynens.be/notes/mysql-utf8mb4. Of note in that article is
the disclaimer that the MySQL client and server settings have to be
updated to use the new character set and collation, something which
cannot be performed in a migration.

String fields with indexes defined on them had to be narrowed to 191
characters due to an inadequacy of the InnoDB database engine. This
operation was only performed on fields that actually needed indexes; a
number of indexes that were determined to not actually be used were
removed.

👌