Added deploy hooks to deploy:migrations
The deploy script has hooks that must be run for the email pulling
service to function properly. Previously, these hooks were only
installed on the deploy task, which meant that in deploys where
migrations need to be run (so, the deploy:migrations task would be run
instead) the email service would not function properly.