Event form no longer always adds empty role to subevents
The event edit form adds an empty role to subevents so that the form
doesn't look strangely empty, but this is unnecessary if there are
already roles on that subevent. This also fixes the problem caused by
the fact that people who have permission to add assistant roles but not
all roles don't have the blank option in the role type list, which
basically caused an unassigned assistant position to be added to the
subevent every time the form was loaded, and if this went unnoticed by
the editor, could cause multiple assistant positions to be added to the
subevent, which would all be appliable by default.