Bring attention to when call/strike dates differ from event date.
On the main events list page, add additional data under "Timing" to bring
attention to "event dates" with elements that span multiple calendar days,
whether intentionally or accidentally.  In particular, since the "Date" column
shows the calendar date for event start, this is important if call, event end,
and/or strike are on a different calendar day.

There is an exception for event end and strike in the typical edge case of a
several-hour event running somewhat past midnight, in which case the intent is
relatively straightforward without the extra notation.  For longer events, it
is probably more helpful to include anyway.