Feat: Allow Invoice Items to have Notes (#604)
Historically, "invoice items" were the presets that could be applied when creating an invoice. They've fell out of favor, but we still generate massively templated invoices.

This PR adds "notes" to invoice items, such that when a preset is used, the memo is autofilled. We regularly use memos like this for labor quoting and other packages.

Co-authored-by: Perry Naseck <4472083+DaAwesomeP@users.noreply.github.com>